+++ src/components/ui/theme/ThemeInterface.ts
@@ -7,6 +7,18 @@ export interface ThemeInterface {
     secondaryText: string;
     highlight: string;
     highlightText: string;
+    conditional: {
+      level1: string;
+      level2: string;
+      level3: string;
+      level4: string;
+    };
+    token: {
+      variable: string;
+      character: string;
+      speed: string;
+      text: string;
+    }
     toolbar: {
       background: string;
       inactiveBackground: string;
@@ -43,14 +55,14 @@ export interface ThemeInterface {
     sidebar: {
       background: string;
       border: string;
-    }
+    };
     document: {
       background: string;
-    },
+    };
     list: {
       selectedBackground: string;
       activeBackground: string;
-    }
+    };
   };
   borderRadius: number;
 }